Created a formatTime function and called the function.
To format the current time into 12 hour Standard Time. Applying the padStart() method on the hour and minute strings with the following parameters: 2 specifies the minimum length of the resulting string. If the length is less than 2, leading zeroes will be added. "0" specifies the character to be used for padding (leading zeroes, in this case).